requirements.txt
discord.py
google-cloud-vision
google-auth
asyncio
